data-pjt/producer/requirements.txt
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.13.3
python-dotenv==1.0.1
kafka-python==2.0.3
psycopg2-binary==2.9.9
